refactor(front-end): Fix React hooks deps in Template
