+++ Source/Plugins/Core/com.equella.core/js/tsrc/mainui/MainMenu.tsx
@@ -1,0 +1,97 @@
+/*
+ * Licensed to The Apereo Foundation under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * The Apereo Foundation licenses this file to you under the Apache License,
+ * Version 2.0, (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import {
+  Divider,
+  Icon,
+  List,
+  ListItem,
+  ListItemIcon,
+  ListItemText,
+  Typography,
+  useTheme,
+} from "@material-ui/core";
+import * as OEQ from "@openequella/rest-api-client";
+import * as React from "react";
+import { Link } from "react-router-dom";
+import { useStyles } from "./Template";
+
+interface MainMenuProps {
+  menuGroups: Array<Array<OEQ.LegacyContent.MenuItem>>;
+  onClickNavItem: () => void;
+}
+
+const MainMenu = ({ menuGroups, onClickNavItem }: MainMenuProps) => {
+  const classes = useStyles(useTheme());
+
+  const navItem = (item: OEQ.LegacyContent.MenuItem, ind: number) => (
+    <ListItem
+      component={(p) => {
+        const props = {
+          ...p,
+          rel: item.newWindow ? "noopener noreferrer" : undefined,
+          target: item.newWindow ? "_blank" : undefined,
+          onClick: onClickNavItem,
+        };
+        return item.route ? (
+          <Link {...props} to={item.route} />
+        ) : (
+          <a {...props} href={item.href}>
+            {}
+          </a>
+        );
+      }}
+      key={ind}
+      button
+    >
+      <ListItemIcon>
+        {item.iconUrl ? (
+          <img src={item.iconUrl} alt={item.title} />
+        ) : (
+          <Icon color="inherit" className={classes.menuIcon}>
+            {item.systemIcon ? item.systemIcon : "folder"}
+          </Icon>
+        )}
+      </ListItemIcon>
+      <ListItemText
+        disableTypography
+        primary={
+          <Typography
+            variant="subtitle1"
+            className={classes.menuItem}
+            component="div"
+          >
+            {item.title}
+          </Typography>
+        }
+      />
+    </ListItem>
+  );
+
+  return (
+    <div id="menulinks">
+      {menuGroups.map((group, ind) => (
+        <React.Fragment key={ind}>
+          {ind > 0 && <Divider />}
+          <List component="nav">{group.map(navItem)}</List>
+        </React.Fragment>
+      ))}
+    </div>
+  );
+};
+
+export default React.memo(MainMenu);
